+++ Vape-X/ValideteViewController.swift
@@ -1,0 +1,8 @@
+//
+//  ValideteViewController.swift
+//  Vape-X
+//
+//  Created by Bilal Naeem on 10/5/22.
+//
+
+import Foundation
